Move `SessionCallOutcome` to `linera-base`
Replace the duplicated types in `linera-execution` and `linera-sdk`, and
update the fields of the WIT type.
